services/processing-service/pyproject.toml
[project]
name = "processing-service"
version = "0.1.0"
description = "Core Domain do ArchLens — consome diagramas, analisa via LLM (vision), identifica componentes e riscos arquiteturais."
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "aio-pika>=9.5.0",
    "aioboto3>=13.0.0",
    "pydantic-ai>=0.1.0",
    "PyMuPDF>=1.25.0",
    "Pillow>=11.0.0",
    "tenacity>=9.0.0",
    "pybreaker>=1.2.0",
    "newrelic>=10.0.0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = ["asyncio"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pre-commit>=4.5.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.9",
    "ty>=0.0.26",
]
